xaynet-sdk: move unittests to their own directory
