+++ rust/xaynet-sdk/src/state_machine/tests/mod.rs
@@ -1,0 +1,2 @@
+mod phases;
+pub mod utils;
